merging

+++ includes/libft/src/ft_strndup.c
@@ -6,7 +6,7 @@
 /*   By: joaocard <joaocard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/18 12:43:55 by joaocard          #+#    #+#             */
-/*   Updated: 2024/01/18 12:50:59 by joaocard         ###   ########.fr       */
+/*   Updated: 2024/01/18 15:50:48 by joaocard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ char	*ft_strndup(const char *s, size_t n)
 {
 	size_t	len;
 	char	*new;
-	int		i;
+	size_t	i;
 	len = 0;
 	i = 0;
 	while (len < n && s[len])
